Addressing review comments and correcting tests.

+++ src/test/java/uk/co/ramp/OutbreakTest.java
@@ -153,7 +153,7 @@ public class OutbreakTest {
             .filter(status -> status == SUSCEPTIBLE)
             .count();
 
-    Map<Double, CmptRecord> records = outbreak.propagate();
+    Map<Integer, CmptRecord> records = outbreak.propagate();
 
     long susceptiblePost =
         population.values().stream()
